[solvers] Hoist semidefinite relaxation code from anon to internal
