+++ solvers/BUILD.bazel
@@ -1080,6 +1080,9 @@ drake_cc_library(
         ":mathematical_program",
         "//math:matrix_util",
     ],
+    deps = [
+        ":semidefinite_relaxation_internal",
+    ],
 )
 
 drake_cc_library(
