+++ solvers/semidefinite_relaxation_internal.h
@@ -1,11 +1,47 @@
 #pragma once
 
+#include <map>
+
 #include "drake/math/matrix_util.h"
 #include "drake/solvers/mathematical_program.h"
 
 namespace drake {
 namespace solvers {
 namespace internal {
+
+void ValidateProgramIsSupported(const MathematicalProgram& prog);
+
+// Check whether the program prog has any non-convex quadratic costs or
+// constraints.
+bool CheckProgramHasNonConvexQuadratics(const MathematicalProgram& prog);
+
+// Iterate over the quadratic costs and constraints in prog, remove them if
+// present in the relaxation, and add an equivalent linear cost or constraint on
+// the semidefinite variable X. The map variables_to_sorted_indices maps the
+// decision variables in prog to their index in the last column of X.
+void DoLinearizeQuadraticCostsAndConstraints(
+    const MathematicalProgram& prog, const MatrixXDecisionVariable& X,
+    const std::map<symbolic::Variable, int>& variables_to_sorted_indices,
+    MathematicalProgram* relaxation);
+
+// Aggregate all the finite linear constraints in the program into a single
+// expression Ay ≤ b, which can be expressed as [A, -b][y; 1] ≤ 0.
+// We add the implied linear constraint [A,-b]X[A,-b]ᵀ ≤ 0 on the variable X to
+// the relaxation. The map variables_to_sorted_indices maps the
+// decision variables in prog to their index in the last column of X.
+void DoAddImpliedLinearConstraints(
+    const MathematicalProgram& prog, const MatrixXDecisionVariable& X,
+    const std::map<symbolic::Variable, int>& variables_to_sorted_indices,
+    MathematicalProgram* relaxation);
+
+// For every equality constraint Ay = b in prog, add the implied linear equality
+// constraint [A, -b]X = 0 on the semidefinite relaxation variable X to the
+// relaxation. The map variables_to_sorted_indices maps the decision variables
+// in prog to their index in the last column of X.
+void DoAddImpliedLinearEqualityConstraints(
+    const MathematicalProgram& prog, const MatrixXDecisionVariable& X,
+    const std::map<symbolic::Variable, int>& variables_to_sorted_indices,
+    MathematicalProgram* relaxation);
 
 // Take the sparse Kronecker product of A and B.
 Eigen::SparseMatrix<double> SparseKroneckerProduct(
